Added check to detect and exclude follower loops
Before, loop conditions were possible if a follower was visited twice. This was caused by a user following someone and that person following the user as well or  down the line.

Now, each follower's name is checked to see if it has already been visited. If it has, it is not appended to the list, therefore not visited.

Fixes issue #4 in regards to the followers API endpoint.