Fixed loop issue in repos API
Before, loop conditions were possible if a stargazer was visited twice. This was caused by a stargazer staring the same repository a previously visited stargazer stared.

Now, each stargazer's name is checked to see if it has already been visited. If it has, it is not appended to the list of users, and therefore not visited.

Same logic was used to fix the looping issue in the followers endpoint.

Fixes issue #4